+++ 101_FT_VEC/ft_vec1.c
@@ -12,19 +12,19 @@
 
 #include "ft_vec.h"
 
-inline float	ft_vec_dot(const t_vec v1, const t_vec v2)
+inline double	ft_vec_dot(const t_vec v1, const t_vec v2)
 {
 	return ((*v1 * *v2) + (*(v1 + 1) * *(v2 + 1)) + (*(v1 + 2) * *(v2 + 2)));
 }
 
-inline float	ft_vec_mag(const t_vec v)
+inline double	ft_vec_mag(const t_vec v)
 {
-	return (sqrtf(ft_vec_dot(v, v)));
+	return (sqrt(ft_vec_dot(v, v)));
 }
 
 void	ft_vec_norm(t_vec *newv, const t_vec v)
 {
-	const float	mag = ft_vec_mag(v);
+	const double	mag = ft_vec_mag(v);
 
 	if (mag > 0)
 		ft_vec_div(newv, v, (t_vec){mag, mag, mag});
